Introduce `#fill_in` System Tests
This commit introduces the `WithForm::TestHelpers#with_form` method,
which yields instances of the new `WithForm::Form` class.

Instances of the `WithForm::Form` class expose the `#fill_in` and
`#label` methods.

The `#fill_in` method mimics the [interface of the Capybara-provided
method of the same name][fill_in], except that it accepts an
internationalization scope key as its first argument, which it passes to
the `#label` method to resolve the corresponding `<label>` element's
text.

In support of test-driving these methods, this commit adds dependencies
on the [Capybara gem][capybara], along with the [Webdrivers
gem][webdrivers] and the [Selenium-Webdriver gem][selenium-webdriver].

[fill_in]: https://github.com/teamcapybara/capybara#interacting-with-forms
[capybara]: https://github.com/teamcapybara/capybara
[webdrivers]: https://github.com/titusfortner/webdrivers
[selenium-webdriver]: https://github.com/SeleniumHQ/selenium