Add `#attach_file` test helper
Implements the `WithForm::Form#attach_file` method to mimic Capybara's
[`#attach_file`][attach_file] method.

To test the process, add the `test/fixtures/file.txt` file, and "upload"
it via the
[`ActiveSupport::Testing::FileFixtures#file_fixture`][file_fixture]
test helper method.

The test's verification phase asserts that the `<input type="file">`
element's value will be `C:\fakepath\file.txt`. That value was
determined by a failed assertion to the contrary. I'm not sure if that
value is predictable or platform independent, but it's an assertion that
passes at the moment, so we'll stick with it for now.

[attach_file]: https://rubydoc.info/github/teamcapybara/capybara/master/Capybara/Node/Actions#attach_file-instance_method
[file_fixture]: https://api.rubyonrails.org/classes/ActiveSupport/Testing/FileFixtures.html#method-i-file_fixture