Introduce `WithForm::ModelForm`
When submitting with an instance of an `ActiveModel::Model`, the
`with_form` helper method will return an instance of
`WithForm::ModelForm` instead of the pre-existing `WithForm::ScopeForm`.

The `model:` option passed to the `WithForm::ModelForm` will be used to
determine its [underlying `scope:` option][scope], which will be used to
create an instance of the `WithForm::ScopeForm` that the `ModelForm`
will delegate calls to.

Unless they specify a value for `with:`, calls to `fill_in` will
determine the value by delegating to the `model:` option instance.

[scope]: https://api.rubyonrails.org/classes/ActiveModel/Naming.html