Add `WithForm::ModelForm#{unselct,select}`
This commit adds `WithForm::ModelForm` implementation for the
`WithForm::ScopeForm#{unselect,select}` counterparts.

When invoked with a Symbol for `attribute`, make the selection or
unselection based on the `model:` option's value for the corresponding
attribute.

When the `from:` option is specified, forward it along to the underlying
`ScopeForm` instance. Otherwise, it is an optional argument.