Install Webpacker into the Dummy application
The [`rails webpacker:install`][webpacker:install] implementation
assumes that it will be invoked from a "Real Host Application", so when
executing the task within an engine's Dummy application, it doesn't
prefix the tasks with `app:`. In this case, the properly named task
would coincidentally be prefixed with `app:` twice: `app:app:template`.

To circumvent that issue, this commit was generated by executing these
two commands, in sequence:

```bash

touch test/dummy/.gitignore
rails app:app:template \
  LOCATION="$(bundle info --path webpacker)/lib/install/template.rb"
```

The first command is necessary, since the Webpacker template assumes
the existence of `.gitignore`.

The second command is lifted from the [`webpacker:install` task
definition][webpacker:install-template], using [`bundle info
--path`][bundle-info] as a means of dynamically resolving the full
system path to the Webpacker gem's `template.rb` definition.

In a ddition to the generated contents of this file, this commit adds
the root level `node_modules/` directory to the project's `.gitignore`.

[webpacker:install]: https://github.com/rails/webpacker/blob/v5.0.1/lib/tasks/installers.rake#L26-L30
[webpacker:install-template]: https://github.com/rails/webpacker/blob/v5.0.1/lib/tasks/installers.rake#L24-L38
[bundle-info]: https://bundler.io/man/bundle-info.1.html#OPTIONS