Improve `type="checkbox"` tests and guides
This commit improves support for `check` and `uncheck` calls to
`with_form(model:)` invocations that are passed `ActiveRecord::Base`
instances.

To support this, add migrations to add a `Boolean` column to the
`widget_records` table, a new `Tag` model and table to be associated
to `Widget` records through the `Tagging` model and table.