Extract `WithForm::ModelForm#read_attribute`
When `with_form(model:)` instances are passed a [`Symbol`][ruby-symbol],
they delegate a method with a matching name to their underlying models,
and pass the value that's returned along to a `WithForm::ScopeForm`
instance.

There was structural and conceptual duplication of that guarding and
reading logic spread across several method implementations.

This commit extracts that to a private `read_attribute` method, and
re-structures the call sites.

[ruby-symbol]: https://ruby-doc.org/core-2.7.1/Symbol.html